Rename prompt to dialog
