+++ src/cmd/append.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/git-town/git-town/v7/src/cli"
 	"github.com/git-town/git-town/v7/src/git"
-	"github.com/git-town/git-town/v7/src/prompt"
+	"github.com/git-town/git-town/v7/src/dialog"
 	"github.com/git-town/git-town/v7/src/runstate"
 	"github.com/spf13/cobra"
 )
@@ -79,7 +79,7 @@ func createAppendConfig(args []string, repo *git.ProdRepo) (result appendConfig,
 	if hasBranch {
 		return result, fmt.Errorf("a branch named %q already exists", result.targetBranch)
 	}
-	err = prompt.EnsureKnowsParentBranches([]string{result.parentBranch}, repo)
+	err = dialog.EnsureKnowsParentBranches([]string{result.parentBranch}, repo)
 	if err != nil {
 		return result, err
 	}
