+++ src/cmd/diff_parent.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/git-town/git-town/v7/src/cli"
 	"github.com/git-town/git-town/v7/src/git"
-	"github.com/git-town/git-town/v7/src/prompt"
+	"github.com/git-town/git-town/v7/src/dialog"
 	"github.com/spf13/cobra"
 )
 
@@ -64,7 +64,7 @@ func createDiffParentConfig(args []string, repo *git.ProdRepo) (config diffParen
 	if !prodRepo.Config.IsFeatureBranch(config.branch) {
 		return config, fmt.Errorf("you can only diff-parent feature branches")
 	}
-	err = prompt.EnsureKnowsParentBranches([]string{config.branch}, repo)
+	err = dialog.EnsureKnowsParentBranches([]string{config.branch}, repo)
 	if err != nil {
 		return config, err
 	}
