+++ src/cmd/hack.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/git-town/git-town/v7/src/cli"
 	"github.com/git-town/git-town/v7/src/git"
-	"github.com/git-town/git-town/v7/src/prompt"
+	"github.com/git-town/git-town/v7/src/dialog"
 	"github.com/git-town/git-town/v7/src/runstate"
 	"github.com/spf13/cobra"
 )
@@ -50,11 +50,11 @@ See "sync" for information regarding remote upstream.`,
 
 func determineParentBranch(targetBranch string, repo *git.ProdRepo) (string, error) {
 	if promptForParent {
-		parentBranch, err := prompt.AskForBranchParent(targetBranch, repo.Config.MainBranch(), repo)
+		parentBranch, err := dialog.AskForBranchParent(targetBranch, repo.Config.MainBranch(), repo)
 		if err != nil {
 			return "", err
 		}
-		err = prompt.EnsureKnowsParentBranches([]string{parentBranch}, repo)
+		err = dialog.EnsureKnowsParentBranches([]string{parentBranch}, repo)
 		if err != nil {
 			return "", err
 		}
