+++ src/dialog/branch_helpers.go
@@ -1,4 +1,4 @@
-package prompt
+package dialog
 
 import (
 	"fmt"
