+++ src/dialog/doc.go
@@ -1,0 +1,2 @@
+// Package dialog provides CLI-based dialogs through which the user can enter configuration data.
+package dialog
