+++ src/dialog/parent_branches.go
@@ -1,4 +1,4 @@
-package prompt
+package dialog
 
 import (
 	"fmt"
