+++ src/dialog/unfinished_run_state.go
@@ -1,4 +1,4 @@
-package prompt
+package dialog
 
 import (
 	"fmt"
