--- src/prompt/doc.go
@@ -1,2 +1,0 @@
-// Package prompt provides CLI-based dialogs through which the user can enter configuration data.
-package prompt
