+++ src/steps/squash_merge_branch_step.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/git-town/git-town/v7/src/git"
 	"github.com/git-town/git-town/v7/src/hosting"
-	"github.com/git-town/git-town/v7/src/prompt"
+	"github.com/git-town/git-town/v7/src/dialog"
 )
 
 // SquashMergeBranchStep squash merges the branch with the given name into the current branch.
@@ -36,7 +36,7 @@ func (step *SquashMergeBranchStep) Run(repo *git.ProdRepo, driver hosting.Driver
 	if err != nil {
 		return err
 	}
-	author, err := prompt.DetermineSquashCommitAuthor(step.BranchName, repo)
+	author, err := dialog.DetermineSquashCommitAuthor(step.BranchName, repo)
 	if err != nil {
 		return fmt.Errorf("error getting squash commit author: %w", err)
 	}
